fix: 피해량을 float32로 떨어뜨린 뒤 올린다
게임은 피해를 float32로 굴리는데 우리는 double로 재서, 값이 정수에 아주 가까울 때
우리 쪽 찌꺼기 때문에 한 칸 더 올라갔다.
  현령 공명 해방 크리 — double 93438.0009 → 93439 / float32 93438.0000 → 93438(게임값)
같은 실측의 나머지 아홉 줄은 올림까지 0.009 넘게 남아 있어 그대로다.

+++ src/calculator/anomaly.ts
@@ -50,7 +50,8 @@ function anomalyDefMultiplier(charLevel: number, enemyLevel: number, defReductio
   return numerator / (numerator + enemyDef * (1 - defReduction));
 }
 
-const ceilDamage = (v: number) => Math.ceil(v - 1e-9);
+/** 피해량 올림. 게임과 같은 자리에서 올리도록 float32로 떨어뜨린 뒤 올린다(damage.ts 설명 참고). */
+const ceilDamage = (v: number) => Math.ceil(Math.fround(v) - 1e-9);
 
 export interface AnomalyDamageInput {
   kind: AnomalyKind;
